refactor: modularized edit documents on case and homologued styles
